remove logic from model
